reorganized files to match the folder structure from PR #510
