+++ frontend/src/components/Leagues/GlobalRanking/GlobalRanking.tsx
