+++ frontend/src/components/Leagues/LeaderCard/LeaderCard.tsx
