+++ frontend/src/components/Leagues/WeeklyRanking/WeeklyRanking.tsx
@@ -1,14 +1,14 @@
 import { useEffect, useState } from "react";
-import { getLeagueRanking } from "../../services/leagueService";
-import type { Liga } from "../../types/league";
-import { StandingsTable } from "../leagues/StandingsTable";
+import { getLeagueRanking } from "../../../services/leagueService";
+import type { Liga } from "../../../types/league";
+import { StandingsTable } from "../StandingsTable/StandingsTable";
 
 export const WeeklyRanking = () => {
   const [data, setData] = useState<Liga[]>([]);
   useEffect(() => {
     getLeagueRanking()
       .then(setData)
-      .catch(() => {});
+      .catch(() => { });
   }, []);
 
   return (
